Rename plugins.py to pluginmanager.py
This is to avoid a name conflict with documentation.
