+++ nbodykit/pluginmanager.py
